+++ cpp/lib/src/link/LinkContext.cpp
@@ -50,6 +50,16 @@ LinkContext::LinkContext(const Logger& logger,
 {
 }
 
+std::shared_ptr<LinkContext> LinkContext::Create(const Logger& logger,
+                                                 const std::shared_ptr<exe4cpp::IExecutor>& executor,
+                                                 std::shared_ptr<IUpperLayer> upper,
+                                                 std::shared_ptr<ILinkListener> listener,
+                                                 ILinkSession& session,
+                                                 const LinkLayerConfig& config)
+{
+    return std::shared_ptr<LinkContext>(new LinkContext(logger, executor, upper, listener, session, config));
+}
+
 bool LinkContext::OnLowerLayerUp()
 {
     if (this->isOnline)
@@ -58,13 +68,9 @@ bool LinkContext::OnLowerLayerUp()
         return false;
     }
 
-    const auto now = Timestamp(this->executor->get_time());
-
     this->isOnline = true;
 
-    this->lastMessageTimestamp = now; // no reason to trigger a keep-alive until we've actually expired
-
-    this->StartKeepAliveTimer(now + config.KeepAliveTimeout);
+    this->RestartKeepAliveTimer();
 
     listener->OnStateChange(LinkStatus::UNRESET);
     upper->OnLowerLayerUp();
@@ -230,16 +236,14 @@ void LinkContext::TryStartTransmission()
 void LinkContext::OnKeepAliveTimeout()
 {
     const auto now = Timestamp(this->executor->get_time());
-
-    auto elapsed = now - this->lastMessageTimestamp;
+    const auto elapsed = now - this->lastMessageTimestamp;
 
     if (elapsed >= this->config.KeepAliveTimeout)
     {
-        this->lastMessageTimestamp = now;
         this->keepAliveTimeout = true;
     }
 
-    this->StartKeepAliveTimer(now + config.KeepAliveTimeout);
+    this->RestartKeepAliveTimer();
 
     this->TryStartTransmission();
 }
@@ -253,12 +257,27 @@ void LinkContext::OnResponseTimeout()
 
 void LinkContext::StartResponseTimer()
 {
-    rspTimeoutTimer = executor->start(config.Timeout.value, [this]() { this->OnResponseTimeout(); });
+    this->rspTimeoutTimer = executor->start(config.Timeout.value, [self = shared_from_this()]() {
+        if (self->isOnline)
+        {
+            self->OnResponseTimeout();
+        }
+    });
 }
 
-void LinkContext::StartKeepAliveTimer(const Timestamp& expiration)
+void LinkContext::RestartKeepAliveTimer()
 {
-    this->keepAliveTimer = executor->start(expiration.value, [this]() { this->OnKeepAliveTimeout(); });
+    this->keepAliveTimer.cancel();
+
+    this->lastMessageTimestamp = Timestamp(this->executor->get_time());
+    const auto expiration = this->lastMessageTimestamp + this->config.KeepAliveTimeout;
+
+    this->keepAliveTimer = executor->start(expiration.value, [self = shared_from_this()]() {
+        if (self->isOnline)
+        {
+            self->OnKeepAliveTimeout();
+        }
+    });
 }
 
 void LinkContext::CancelTimer()
@@ -310,31 +329,19 @@ bool LinkContext::OnFrame(const LinkHeaderFields& header, const ser4cpp::rseq_t&
         return false;
     }
 
-    if (header.addresses.IsBroadcast())
+    // Broadcast addresses can only be used for sending data.
+    // If confirmed data is used, no response is sent back.
+    if (header.addresses.IsBroadcast() &&
+        !(header.func == LinkFunction::PRI_UNCONFIRMED_USER_DATA || header.func == LinkFunction::PRI_CONFIRMED_USER_DATA))
     {
-        // Broadcast addresses can only be used for sending data.
-        // If confirmed data is used, no response is sent back.
-        if (header.func == LinkFunction::PRI_UNCONFIRMED_USER_DATA)
-        {
-            this->PushDataUp(Message(header.addresses, userdata));
-            return true;
-        }
-        else if (header.func == LinkFunction::PRI_CONFIRMED_USER_DATA)
-        {
-            pSecState = &pSecState->OnConfirmedUserData(*this, header.addresses.source, header.fcb, true,
-                                                        Message(header.addresses, userdata));
-        }
-        else
-        {
-            FORMAT_LOG_BLOCK(logger, flags::WARN, "Received invalid function (%s) with broadcast destination address",
-                             LinkFunctionSpec::to_string(header.func));
-            ++statistics.numUnexpectedFrame;
-            return false;
-        }
+        FORMAT_LOG_BLOCK(logger, flags::WARN, "Received invalid function (%s) with broadcast destination address",
+                            LinkFunctionSpec::to_string(header.func));
+        ++statistics.numUnexpectedFrame;
+        return false;
     }
 
     // reset the keep-alive timestamp
-    this->lastMessageTimestamp = Timestamp(this->executor->get_time());
+    this->RestartKeepAliveTimer();
 
     switch (header.func)
     {
@@ -360,7 +367,7 @@ bool LinkContext::OnFrame(const LinkHeaderFields& header, const ser4cpp::rseq_t&
         pSecState = &pSecState->OnRequestLinkStatus(*this, header.addresses.source);
         break;
     case (LinkFunction::PRI_CONFIRMED_USER_DATA):
-        pSecState = &pSecState->OnConfirmedUserData(*this, header.addresses.source, header.fcb, false,
+        pSecState = &pSecState->OnConfirmedUserData(*this, header.addresses.source, header.fcb, header.addresses.IsBroadcast(),
                                                     Message(header.addresses, userdata));
         break;
     case (LinkFunction::PRI_UNCONFIRMED_USER_DATA):